backend/requirements.txt
# ─── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# ─── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9
greenlet==3.1.1

# ─── Config / validation ───────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ─── Auth (Clerk JWT verification) ────────────────────────────────────────────
python-jose[cryptography]==3.3.0
httpx==0.27.2

# ─── Task queue ────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.1.1

# ─── Recipe scraping ───────────────────────────────────────────────────────────
recipe-scrapers==15.5.0
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# ─── YouTube ingestion ─────────────────────────────────────────────────────────
yt-dlp>=2025.1.15
faster-whisper>=1.0.3          # replaces openai-whisper; modern packaging, 4× faster

# ─── AI / OpenRouter ───────────────────────────────────────────────────────────
openai==1.51.0
